use int, rather than smallint
git-svn-id: file:///Users/michaelkoepke/svn/semiologic@4254 6c9c31c4-cb3d-4424-af3d-876279169c7e

+++ plugins/sem-backend/trunk/schema/utils.depletable.sql
@@ -18,7 +18,7 @@ BEGIN
 	THEN
 		EXECUTE $EXEC$
 		ALTER TABLE $EXEC$ || quote_ident(t_name) || $EXEC$
-			ADD COLUMN $EXEC$ || quote_ident(t_field) || $EXEC$ smallint;
+			ADD COLUMN $EXEC$ || quote_ident(t_field) || $EXEC$ int;
 		$EXEC$;
 	END IF;
